feat: Interface層: GroupをUserから分離
