+++ api/user/internal/application/request/group.go
@@ -1,0 +1,7 @@
+package request
+
+// CreateGroup - グループ登録APIのリクエスト
+type CreateGroup struct {
+	Name        string `json:"name" validate:"required"`
+	Description string `json:"description"`
+}
