+++ api/user/internal/application/request/user.go
@@ -6,9 +6,3 @@ type CreateUser struct {
 	Password             string `json:"password" validate:"required"`
 	PasswordConfirmation string `json:"passwordConfirmation" validate:"required,eqfield=Password"`
 }
-
-// CreateGroup - グループ登録APIのリクエスト
-type CreateGroup struct {
-	Name        string `json:"name" validate:"required"`
-	Description string `json:"description"`
-}
